llm_enricher/requirements.txt
aiohttp>=3.9.0
asyncpg>=0.29.0
httpx>=0.27.0
tenacity>=8.2.0
pybreaker>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
jsonschema>=4.21.0
pyyaml>=6.0.1
prometheus-client>=0.20.0
pytest>=8.1.0
pytest-asyncio>=0.23.0
respx>=0.21.0
